+++ packages/statemanager/src/rpcStateManager.ts
@@ -358,7 +358,7 @@ export class RPCStateManager implements StateManagerInterface {
   }
 
   // TODO TRON implement token in rpc state manager
-  async tokenIdExists(_tokenId: number): Promise<boolean> {
+  async tokenIdExists(_tokenId: bigint): Promise<boolean> {
     throw EthereumJSErrorWithoutCode('Method not implemented.')
   }
 }