+++ packages/statemanager/src/statefulBinaryTreeStateManager.ts
@@ -787,7 +787,7 @@ export class StatefulBinaryTreeStateManager implements StateManagerInterface {
     await this.flush()
   }
   // TODO TRON implement token in rpc state manager
-  async tokenIdExists(_tokenId: number): Promise<boolean> {
+  async tokenIdExists(_tokenId: bigint): Promise<boolean> {
     throw EthereumJSErrorWithoutCode('Method not implemented.')
   }
 }