+++ packages/util/test/account.spec.ts
@@ -134,7 +134,7 @@ describe('Account', () => {
       balance: '0x42',
       storageRoot: '0x56e81f171bcc55a6ff8345e692c0f86e5b48e01b996cadc001622fb5e363b421',
       codeHash: '0xc5d2461236f7233c927e7db2dcc703c0e500b653ca82273b7bfad8045d85a470',
-      asset: [],
+      asset: {},
       activePermissions: [],
     }
     const account = createAccount(raw)
@@ -143,7 +143,7 @@ describe('Account', () => {
       raw.balance,
       raw.storageRoot,
       raw.codeHash,
-      raw.asset,
+      [],
       raw.activePermissions,
     ] as Input)